refactor(button-group/list/card): optimize CSS
