+++ src/components/button-group/themes/shared/button/button.bootstrap.scss
@@ -8,16 +8,8 @@ $group-item-border-thickness: rem(1px);
     min-width: var(--button-min-width);
 
     [part~='toggle'] {
-        $icon-size: rem(18px);
-
-        --ig-icon-size: #{$icon-size};
-        --igx-icon-size: #{$icon-size};
-
         min-height: $bootstrap-flat-btn-size;
         border-width: $group-item-border-thickness;
-        border-color: var-get($theme, 'item-border-color');
-        color: var-get($theme, 'item-text-color');
-        background: var-get($theme, 'item-background');
         padding-inline: pad-inline(rem(8px), rem(10px), rem(12px));
         padding-block: pad-block(rem(2px), rem(4px), rem(7px));
         min-width: var(--button-min-width);
@@ -27,8 +19,6 @@ $group-item-border-thickness: rem(1px);
 :host(:hover),
 :host(:focus) {
     [part~='toggle'] {
-        color: var-get($theme, 'item-hover-text-color');
-        background: var-get($theme, 'item-hover-background');
         border-color: var-get($theme, 'item-border-color');
     }
 }
@@ -36,30 +26,5 @@ $group-item-border-thickness: rem(1px);
 :host(:focus) {
     [part~='toggle'] {
         box-shadow: 0 0 0 rem(3px) var-get($theme, 'idle-shadow-color');
-    }
-}
-
-:host([selected]) {
-    [part~='toggle'] {
-        color: var-get($theme, 'item-selected-text-color');
-        background: var-get($theme, 'item-selected-background');
-        border-color: var-get($theme, 'item-selected-border-color');
-    }
-}
-
-:host([selected]:hover),
-:host([selected]:focus) {
-    [part~='toggle'] {
-        color: var-get($theme, 'item-selected-text-color');
-        background: var-get($theme, 'item-selected-hover-background');
-    }
-}
-
-:host([disabled]),
-:host(:disabled) {
-    [part~='toggle'] {
-        color: var-get($theme, 'disabled-text-color');
-        background: var-get($theme, 'disabled-background-color');
-        border-color: var-get($theme, 'item-disabled-border');
     }
 }
