+++ src/components/button-group/themes/shared/button/button.fluent.scss
@@ -13,43 +13,7 @@ $group-item-border-thickness: rem(1px);
 
         min-height: $fluent-flat-btn-size;
         border-width: $group-item-border-thickness;
-        border-color: var-get($theme, 'item-border-color');
-        color: var-get($theme, 'item-text-color');
-        background: var-get($theme, 'item-background');
         padding-inline: pad-inline(rem(16px));
         padding-block: pad-block(rem(7px));
-    }
-}
-
-:host(:hover),
-:host(:focus) {
-    [part~='toggle'] {
-        color: var-get($theme, 'item-hover-text-color');
-        background: var-get($theme, 'item-hover-background');
-    }
-}
-
-:host([selected]) {
-    [part~='toggle'] {
-        color: var-get($theme, 'item-selected-text-color');
-        background: var-get($theme, 'item-selected-background');
-        border-color: var-get($theme, 'item-selected-border-color');
-    }
-}
-
-:host([selected]:hover),
-:host([selected]:focus) {
-    [part~='toggle'] {
-        color: var-get($theme, 'item-selected-text-color');
-        background: var-get($theme, 'item-selected-hover-background');
-    }
-}
-
-:host([disabled]),
-:host(:disabled) {
-    [part~='toggle'] {
-        color: var-get($theme, 'disabled-text-color');
-        background: var-get($theme, 'disabled-background-color');
-        border-color: var-get($theme, 'item-disabled-border');
     }
 }
