+++ src/components/button-group/themes/shared/button/button.indigo.scss
@@ -6,26 +6,10 @@ $group-item-border-thickness: rem(1px);
 
 :host {
     [part~='toggle'] {
-        $icon-size: rem(18px);
-
-        --ig-icon-size: #{$icon-size};
-        --igx-icon-size: #{$icon-size};
-
         min-height: $indigo-flat-btn-size;
         border-width: $group-item-border-thickness;
-        border-color: var-get($theme, 'item-border-color');
-        color: var-get($theme, 'item-text-color');
-        background: var-get($theme, 'item-background');
         padding-inline: pad-inline(rem(8px), rem(12px), rem(12px));
         padding-block: pad-block(rem(1px), rem(2px), rem(3px));
-    }
-}
-
-:host(:hover),
-:host(:focus) {
-    [part~='toggle'] {
-        color: var-get($theme, 'item-hover-text-color');
-        background: var-get($theme, 'item-hover-background');
     }
 }
 
@@ -37,33 +21,8 @@ $group-item-border-thickness: rem(1px);
     }
 }
 
-:host([selected]) {
-    [part~='toggle'] {
-        color: var-get($theme, 'item-selected-text-color');
-        background: var-get($theme, 'item-selected-background');
-        border-color: var-get($theme, 'item-selected-border-color');
-    }
-}
-
-:host([selected]:hover),
-:host([selected]:focus) {
-    [part~='toggle'] {
-        color: var-get($theme, 'item-selected-text-color');
-        background: var-get($theme, 'item-selected-hover-background');
-    }
-}
-
 :host([selected]:focus) {
     [part~='toggle'] {
         box-shadow: 0 0 0 rem(3px) var-get($theme, 'selected-shadow-color');
-    }
-}
-
-:host([disabled]),
-:host(:disabled) {
-    [part~='toggle'] {
-        color: var-get($theme, 'disabled-text-color');
-        background: var-get($theme, 'disabled-background-color');
-        border-color: var-get($theme, 'item-disabled-border');
     }
 }
