+++ src/components/button-group/themes/shared/group/group.bootstrap.scss
@@ -7,16 +7,3 @@ $theme: $bootstrap;
     --component-size: var(--ig-size, #{var-get($theme, 'default-size')});
     --button-min-width: #{rem(42px)};
 }
-
-[part~='group'] {
-    box-shadow: var-get($theme, 'elevation');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-:host([alignment='horizontal']) {
-    --horizontal-radius: #{var-get($theme, 'border-radius')};
-}
-
-:host([alignment='vertical']) {
-    --vertical-radius: #{var-get($theme, 'border-radius')};
-}
