+++ src/components/button-group/themes/shared/group/group.fluent.scss
@@ -5,18 +5,4 @@ $theme: $fluent;
 
 :host {
     --component-size: var(--ig-size, #{var-get($theme, 'default-size')});
-    --button-min-width: #{rem(50px)};
-}
-
-[part~='group'] {
-    box-shadow: var-get($theme, 'elevation');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-:host([alignment='horizontal']) {
-    --horizontal-radius: #{var-get($theme, 'border-radius')};
-}
-
-:host([alignment='vertical']) {
-    --vertical-radius: #{var-get($theme, 'border-radius')};
 }
