--- src/components/card/themes/shared/container/card.bootstrap.scss
@@ -1,19 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $bootstrap;
-
-:host {
-    border: rem(1px) solid var-get($theme, 'outline-color');
-    background: var-get($theme, 'background');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-:host([elevated]) {
-    box-shadow: var-get($theme, 'resting-elevation');
-}
-
-:host([elevated]:hover),
-:host([elevated]:focus-within) {
-    box-shadow: var-get($theme, 'hover-elevation');
-}
