+++ src/components/card/themes/shared/container/card.fluent.scss
@@ -3,20 +3,8 @@
 
 $theme: $fluent;
 
-:host {
-    border: rem(1px) solid var-get($theme, 'outline-color');
-    background: var-get($theme, 'background');
-    border-radius: var-get($theme, 'border-radius');
-}
-
-:host([elevated]) {
-    box-shadow: var-get($theme, 'resting-elevation');
-}
-
 :host([elevated]:hover),
 :host([elevated]:focus-within) {
-    box-shadow: var-get($theme, 'hover-elevation');
-
     &::after {
         position: absolute;
         content: '';
