+++ src/components/card/themes/shared/content/card.content.bootstrap.scss
@@ -1,12 +1,7 @@
 @use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $bootstrap;
 
 :host {
     @include type-style('body-1');
-
-    color: var-get($theme, 'content-text-color');
 }
 
 ::slotted(*) {
