--- src/components/card/themes/shared/content/card.content.fluent.scss
@@ -1,8 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $fluent;
-
-:host {
-    color: var-get($theme, 'content-text-color');
-}
