--- src/components/card/themes/shared/content/card.content.indigo.scss
@@ -1,8 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $indigo;
-
-:host {
-    color: var-get($theme, 'content-text-color');
-}
