+++ src/components/card/themes/shared/header/card.header.common.scss
