+++ src/components/card/themes/shared/header/card.header.fluent.scss
@@ -1,11 +1,4 @@
 @use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $fluent;
-
-:host {
-    color: var-get($theme, 'header-text-color');
-}
 
 ::slotted([slot='title']) {
     @include type-style('subtitle-1');
@@ -13,6 +6,4 @@ $theme: $fluent;
 
 ::slotted([slot='subtitle']) {
     @include type-style('body-2');
-
-    color: var-get($theme, 'subtitle-text-color');
 }
