--- src/components/card/themes/shared/header/card.header.indigo.scss
@@ -1,12 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $indigo;
-
-:host {
-    color: var-get($theme, 'header-text-color');
-}
-
-::slotted([slot='subtitle']) {
-   color: var-get($theme, 'subtitle-text-color');
-}
