+++ src/components/list/themes/item.ts
@@ -12,38 +12,38 @@ import { styles as indigoLight } from './light/item/list-item.indigo.css.js';
 import { styles as materialLight } from './light/item/list-item.material.css.js';
 // Shared Styles
 import { styles as bootstrap } from './shared/item/list-item.bootstrap.css.js';
+import { styles as shared } from './shared/item/list-item.common.css.js';
 import { styles as fluent } from './shared/item/list-item.fluent.css.js';
 import { styles as indigo } from './shared/item/list-item.indigo.css.js';
-import { styles as material } from './shared/item/list-item.material.css.js';
 import { Themes } from '../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapLight}
+    ${shared} ${bootstrap} ${bootstrapLight}
   `,
   material: css`
-    ${material} ${materialLight}
+    ${shared} ${materialLight}
   `,
   fluent: css`
-    ${fluent} ${fluentLight}
+    ${shared} ${fluent} ${fluentLight}
   `,
   indigo: css`
-    ${indigo} ${indigoLight}
+    ${shared} ${indigo} ${indigoLight}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapDark}
+    ${shared} ${bootstrap} ${bootstrapDark}
   `,
   material: css`
-    ${material} ${materialDark}
+    ${shared} ${materialDark}
   `,
   fluent: css`
-    ${fluent} ${fluentDark}
+    ${shared} ${fluent} ${fluentDark}
   `,
   indigo: css`
-    ${indigo} ${indigoDark}
+    ${shared} ${indigo} ${indigoDark}
   `,
 };
