--- src/components/list/themes/shared/header/list-header.bootstrap.scss
@@ -1,9 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $bootstrap;
-
-:host {
-    background: var-get($theme, 'header-background');
-    color: var-get($theme, 'header-text-color');
-}
