+++ src/components/list/themes/shared/header/list-header.common.scss
