--- src/components/list/themes/shared/header/list-header.indigo.scss
@@ -1,9 +1,0 @@
-@use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $indigo;
-
-:host {
-    background: var-get($theme, 'header-background');
-    color: var-get($theme, 'header-text-color');
-}
