+++ src/components/list/themes/shared/item/list-item.fluent.scss
@@ -4,31 +4,13 @@
 $theme: $fluent;
 
 :host {
-    color: var-get($theme, 'item-text-color');
-    background: var-get($theme, 'item-background');
     border-bottom: rem(1px) solid var-get($theme, 'border-color');
-    border-radius: var-get($theme, 'item-border-radius');
-}
-
-:host(:hover),
-:host(:focus-within) {
-    background: var-get($theme, 'item-background-hover');
 }
 
 :host(:last-of-type) {
     border-bottom: none;
 }
 
-::slotted(igc-icon) {
-    color: var-get($theme, 'item-action-color');
-}
-
 [part='title'] {
     @include type-style('caption');
-
-    color: var-get($theme, 'item-title-color');
-}
-
-[part='subtitle'] {
-    color: var-get($theme, 'item-subtitle-color');
 }
