+++ src/components/list/themes/shared/item/list-item.indigo.scss
@@ -1,29 +1,5 @@
 @use 'styles/utilities' as *;
-@use '../../light/themes' as *;
-
-$theme: $indigo;
-
-:host {
-    color: var-get($theme, 'item-text-color');
-    background: var-get($theme, 'item-background');
-    border-radius: var-get($theme, 'item-border-radius');
-}
-
-:host(:hover),
-:host(:focus-within) {
-    background: var-get($theme, 'item-background-hover');
-}
-
-::slotted(igc-icon) {
-    color: var-get($theme, 'item-action-color');
-}
-
-[part='title'] {
-    color: var-get($theme, 'item-title-color');
-}
 
 [part='subtitle'] {
     @include type-style('body-2');
-
-    color: var-get($theme, 'item-subtitle-color');
 }
